pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "shieldflow"
version = "0.1.0"
description = "ShieldFlow: LLM Runtime IDS for prompts, responses, and tool calls"
authors = [{ name = "ShieldFlow Team" }]
readme = "README.md"
requires-python = ">=3.9"
keywords = ["llm", "security", "pii", "kafka", "datadog", "runtime-ids"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
observability = ["requests>=2.28"]
redis = ["redis>=4.5"]
langchain = ["langchain>=0.1", "tiktoken>=0.5"]
crewai = ["crewai>=0.28"]
streaming = ["apache-flink==2.2.0", "kafka-python>=2.0"]
test = ["pytest>=7.4", "kafka-python>=2.0", "requests>=2.28", "apache-flink==2.2.0"]

[project.urls]
Homepage = "https://example.com/shieldflow"

[tool.pytest.ini_options]
markers = [
    "integration: tests that require external stack (Redis/Kafka/Flink)",
]
